+++ src/one_dragon_qt/widgets/welcome_dialog.py
@@ -1,19 +1,17 @@
-from PySide6.QtCore import Qt, QTimer, QUrl
-from PySide6.QtGui import QDesktopServices
-from PySide6.QtWidgets import QHBoxLayout, QSpacerItem, QSizePolicy
-from qfluentwidgets import MessageBoxBase, PushButton, SubtitleLabel, DisplayLabel
+from PySide6.QtCore import QTimer
+from qfluentwidgets import MessageBoxBase, SubtitleLabel, DisplayLabel
 
 class WelcomeDialog(MessageBoxBase):
     """首次运行时显示的欢迎对话框"""
 
-    def __init__(self, parent=None):
+    def __init__(self, parent=None, title="欢迎使用一条龙"):
         super().__init__(parent)
 
         self.cancelButton.hide()
         self.yesButton.setText("确定(5s)")
         self.yesButton.setEnabled(False)
 
-        self.titleLabel = SubtitleLabel("欢迎使用绝区零一条龙")
+        self.titleLabel = SubtitleLabel(title)
         self.viewLayout.addWidget(self.titleLabel)
 
         content_label = DisplayLabel(self)
@@ -24,33 +22,8 @@ class WelcomeDialog(MessageBoxBase):
         self._start_countdown()
 
     def _setup_buttons(self):
-        """设置对话框按钮"""
-        quick_start_button = PushButton("快速开始", self)
-        quick_start_button.clicked.connect(lambda: QDesktopServices.openUrl(QUrl("https://onedragon-anything.github.io/zzz/zh/quickstart.html")))
-        quick_start_button.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
-        quick_start_button.adjustSize()
-
-        doc_button = PushButton("自助排障", self)
-        doc_button.clicked.connect(lambda: QDesktopServices.openUrl(QUrl("https://www.kdocs.cn/l/cbSJUUNotJ3Z")))
-        doc_button.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
-        doc_button.adjustSize()
-
-        github_button = PushButton("开源地址", self)
-        github_button.clicked.connect(lambda: QDesktopServices.openUrl(QUrl("https://github.com/OneDragon-Anything/ZenlessZoneZero-OneDragon")))
-        github_button.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
-        github_button.adjustSize()
-
-        spacer = QSpacerItem(10, 0, QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Minimum)
-
-        button_layout = QHBoxLayout()
-        button_layout.addStretch(1)
-        button_layout.addWidget(quick_start_button)
-        button_layout.addItem(spacer)
-        button_layout.addWidget(doc_button)
-        button_layout.addItem(spacer)
-        button_layout.addWidget(github_button)
-        button_layout.addStretch(1)
-        self.viewLayout.addLayout(button_layout)
+        """设置按钮布局 由子类实现"""
+        pass
 
     def _start_countdown(self):
         """启动倒计时"""
